+++ include/ui/UiUtil.h
@@ -18,8 +18,9 @@ limitations under the License.
 */
 #pragma once
 
-#include <stdint.h>   // for int32_t
 #include <wx/defs.h>  // for wxALL, wxGROW
+
+#include <cstdint>  // for int64_t
 
 #include "ScoreboardCommon.h"  // for DEFAULT_BORDER_SIZE
 
@@ -33,7 +34,7 @@ class UiUtil {
  public:
   // If making a new sizer, prefer using a wxGridBagSizer directly (it has a
   // simple constructor, so it doesn't need this helper method at all)
-  static auto sizer(int32_t rows, int32_t columns) -> wxSizer *;
+  static auto sizer(int64_t rows, int64_t columns) -> wxSizer *;
   static void addToGridBag(wxGridBagSizer *sizer, wxWindow *item, int row,
                            int column, int row_span = 1, int column_span = 1,
                            int border_size = DEFAULT_BORDER_SIZE,
