+++ include/util/StringUtil.h
@@ -17,8 +17,9 @@ limitations under the License.
 */
 #pragma once
 
-#include <stdint.h>     // for int32_t
 #include <wx/string.h>  // for wxString
+
+#include <cstdint>  // for int64_t
 
 namespace cszb_scoreboard {
 
@@ -26,7 +27,7 @@ class StringUtil {
  public:
   static auto intToString(int value) -> wxString;
   static auto stringToInt(const wxString &string, int default_value = 0)
-      -> int32_t;
+      -> int64_t;
 };
 
 }  // namespace cszb_scoreboard
