+++ src/ui/UiUtil.cpp
@@ -29,7 +29,7 @@ namespace cszb_scoreboard {
 // Sizers are always transferred to a window, which takes care of cleanup, so we
 // don't bother with unique_ptr here.  But be careful if you ever need a sizer
 // and _don't_ pass it on to a window.
-auto UiUtil::sizer(int32_t rows, int32_t columns) -> wxSizer * {
+auto UiUtil::sizer(int64_t rows, int64_t columns) -> wxSizer * {
   auto *sizer = new wxFlexGridSizer(rows, columns, 0, 0);
   sizer->SetFlexibleDirection(wxBOTH);
   sizer->SetNonFlexibleGrowMode(wxFLEX_GROWMODE_SPECIFIED);
